requirements.txt
aiohttp
aiosignal
async-timeout
attrs
cachetools
certifi
charset-normalizer
db-dtypes
frozenlist
google-api-core[grpc]
google-auth-oauthlib
google-auth
google-cloud-bigquery-storage
google-cloud-bigquery
google-cloud-core
google-crc32c
google-resumable-media
googleapis-common-protos
grpcio-status
grpcio
idna
multidict
numpy
oauthlib
packaging
pandas-gbq
pandas
proto-plus
protobuf
pyarrow
pyasn1-modules
pyasn1
pydantic
pydata-google-auth
python-dateutil
python-dotenv
pytz
requests-oauthlib
requests
rsa
setuptools
six
typing-extensions
urllib3
yarl
curl_cffi
DrissionPage